requirements.txt
blinker==1.9.0
cachelib==0.13.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Flask==2.3.3
Flask-Caching==2.3.1
Flask-Cors==4.0.0
Flask-Login==0.6.3
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==8.1.0
packaging==25.0
protobuf==4.21.12
requests==2.32.4
urllib3==2.5.0
Werkzeug==2.3.7
